refactor: MovieURL 타입 수정
