+++ BoxOffice/Network/MovieURL.swift
@@ -7,31 +7,9 @@
 
 import Foundation
 
-struct MovieURL {
-    enum DateFormat {
-        static var yyyyMMdd = "yyyyMMdd"
-    }
-    
-    private static let kobisURL = "https://kobis.or.kr/kobisopenapi/webservice/rest/boxoffice/searchDailyBoxOfficeList.json?"
-    private static let detailURL = "https://www.kobis.or.kr/kobisopenapi/webservice/rest/movie/searchMovieInfo.json?"
-    
-    static var kobisURLString: String {
-        return "\(self.kobisURL)key=\(Key.movieDataApiKey)&targetDt=\(fetchTodayDate())"
-    }
-    
-    static func detailURLString(movieCode: String) -> String {
-        return "\(self.detailURL)key=\(Key.movieDataApiKey)&movieCd=\(movieCode)"
-    }
-    
-    private static func fetchTodayDate() -> String {
-        let currentDate = Date()
-        let dateFormatter = DateFormatter()
-        dateFormatter.dateFormat = DateFormat.yyyyMMdd
-        
-        guard let yesterday = Calendar.current.date(byAdding: .day, value: -1, to: currentDate) else { return "" }
-        
-        let targetDate = dateFormatter.string(from: yesterday)
-        
-        return targetDate
-    }
+enum MovieURL {
+    static let schema = "https"
+    static let movieHost = "kobis.or.kr"
+    static let boxofficePath = "/kobisopenapi/webservice/rest/boxoffice/searchDailyBoxOfficeList.json"
+    static let movieDetailPath = "/kobisopenapi/webservice/rest/movie/searchMovieInfo.json"
 }
